Renamed GameObjects.move() to update(). Changed delta values from long to int to cut down on calc overhead. This should work for all deltas less than 596 hours (shouldn't be a problem since frames rarely take that long).

+++ src/games/stendhal/client/entity/ActiveEntity.java
@@ -209,7 +209,7 @@ public abstract class ActiveEntity extends Entity {
 	 * @param	delta		The time (in ms) since last call.
 	 */
 	@Override
-	public void update(final long delta) {
+	public void update(final int delta) {
 		super.update(delta);
 
 		if (!stopped()) {
